refactor(reports): extract types to separate module for LOC compliance
Extract CleanupResult, ReportInfo, and ReportTypeDefinition into
report_cleanup_types.py to reduce report_cleanup_service.py from 414
to 378 lines, meeting CI's 400-line single-file limit.

All tests pass (27/27), type checks pass, lint checks pass.
